Rename RequireSession to not use capital letter in file name
